Static analysis: cppcheck + MISRA C:2012 addon, no findings
- tools/static_analysis.sh: analyses common/ and the App/MCAL layers of
  both ECUs with the real HAL/CMSIS/FreeRTOS headers (GCC/ARM macros
  defined so CMSIS does not stop with #error), fails on any finding
- Fixed in code: 12.2, 10.8, 17.7, 11.8 (required) and 13.4, 17.8, 12.1,
  15.4, 2.5 (advisory); cppcheck variableScope and always-false condition
- Five scoped deviations documented in docs/misra_deviations.md
  (15.5, 8.7/2.5 for common APIs, 8.9, 11.4 in MCAL, 21.6/17.1 for the
  debug log)
- Unit tests pass; both firmwares rebuilt and re-verified on the bus
- Report and README updated

+++ common/uds/uds_server.c
@@ -93,10 +93,10 @@ static uint8_t SessionControl(Ctx_t *c)
     (void)Put(c, (uint8_t)(UDS_SID_DIAG_SESSION_CONTROL + UDS_POSITIVE_OFFSET));
     (void)Put(c, session);
     /* sessionParameterRecord: P2 in 1 ms, P2* in 10 ms resolution */
-    (void)Put(c, (uint8_t)(UDS_P2_SERVER_MS >> 8U));
-    (void)Put(c, (uint8_t)(UDS_P2_SERVER_MS & 0xFFU));
-    (void)Put(c, (uint8_t)((UDS_P2STAR_SERVER_MS / 10U) >> 8U));
-    (void)Put(c, (uint8_t)((UDS_P2STAR_SERVER_MS / 10U) & 0xFFU));
+    (void)Put(c, (uint8_t)((uint16_t)UDS_P2_SERVER_MS >> 8U));
+    (void)Put(c, (uint8_t)((uint16_t)UDS_P2_SERVER_MS & 0xFFU));
+    (void)Put(c, (uint8_t)((uint16_t)(UDS_P2STAR_SERVER_MS / 10U) >> 8U));
+    (void)Put(c, (uint8_t)((uint16_t)(UDS_P2STAR_SERVER_MS / 10U) & 0xFFU));
     return 0U;
 }
 
@@ -213,7 +213,7 @@ static uint8_t ReadDid(Ctx_t *c)
         {
             continue;   /* unknown DIDs are skipped; NRC only if none is known */
         }
-        if ((uint32_t)c->respLen + 2U + did->length > c->respMax)
+        if (((uint32_t)c->respLen + 2U + did->length) > c->respMax)
         {
             return UDS_NRC_RESPONSE_TOO_LONG;
         }
